requirements.txt
# -------------------------------
# Core Python Libraries
# -------------------------------
pandas
requests
tqdm

# -------------------------------
# Playwright (for scraping)
# -------------------------------
playwright
# after install, run: python -m playwright install

# -------------------------------
# PDF Processing
# -------------------------------
pdfplumber
PyPDF2
pillow            # required by pdfplumber for image rendering

# -------------------------------
# OCR
# -------------------------------
pytesseract
# NOTE: You must install Tesseract.exe separately on Windows

# -------------------------------
# Database (MySQL)
# -------------------------------
PyMySQL

# -------------------------------
# Redis (Message Queue)
# -------------------------------
redis

# -------------------------------
# ENV File Loader (optional)
# -------------------------------
python-dotenv

# -------------------------------
# Optional: S3/MinIO support
# (uncomment if later needed)
# boto3

pymysql

